pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sepa-xml-generator"
version = "0.1.0"
description = "GUI + CLI tool to generate SEPA pain.001 XML from CSV files"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "Your Name", email = "you@example.com"}]
license = {text = "MIT"}
dependencies = [
  "PySide6>=6.7",
  "lxml>=5",
  "pydantic>=2.6",
  "typer>=0.12",
  "rich>=13",
  "PyYAML>=6",
]

[project.optional-dependencies]
dev = ["pytest>=8", "pytest-qt>=4", "ruff>=0.6", "black>=24", "mypy>=1.10", "pre-commit>=3"]

[project.scripts]
sepa-xml-generator = "sepa_xml_generator.cli:app"

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
strict = true
mypy_path = "src"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
